refactor: also move into entity
