+++ src/common/entity/has_key_value_pairs.py
