+++ src/tests/bdd/steps/handle_response.py
@@ -6,7 +6,7 @@ from behave import then
 from deepdiff import DeepDiff
 from dictdiffer import diff
 
-from common.utils.data_structure.find import find
+from common.entity.find import find
 from tests.test_helpers.print_pygments import print_pygments
 
 STATUS_CODES = {
