+++ src/tests/unit/services/document_service/test_reference_resolve.py
@@ -5,8 +5,8 @@ from unittest import mock
 import pytest
 
 from common.address import Address
+from common.entity.has_key_value_pairs import has_key_value_pairs
 from common.tree.tree_node_serializer import tree_node_to_dict
-from common.utils.data_structure.has_key_value_pairs import has_key_value_pairs
 from enums import REFERENCE_TYPES, SIMOS, Protocols
 from tests.unit.mock_data.mock_blueprint_provider import MockBlueprintProvider
 from tests.unit.mock_data.mock_document_service import get_mock_document_service
